Changing in to byte on index and other things

+++ tornado.cs
@@ -4,8 +4,8 @@ namespace Cyclone {
   public class Tornado {
     private string[] names = {"Dorra", "SuperSlimy", "ImplausibleSantaClaus", "Volcano", "Poop", "Sharknado", "AgentSquidnado"};
     private string[] desc = {"So powerful you can’t see it coming. So powerful you don’t even know about it.", "Goop?", "This one blows your presents away at Christmas.", "Pompeii met a Cyclone.", "Manure! I hate manure!", "Beware! Shark is Death!", "It's dynamic."};
-    int index;
-    public Tornado(int i){
+    byte index;
+    public Tornado(byte i){
       index = i;
     }
     public string Name{
